refactor(updateTicket): unify service that updates ticket in zendesk and hasura for DRY code
