+++ packages/listener-match/src/components/Individual/forwardPublicService.ts
@@ -1,12 +1,7 @@
 import * as yup from "yup";
-import { updateTicket } from "../Zendesk";
-import {
-  getCurrentDate,
-  composeCustomFields,
-  agentSelectionDicio
-} from "../../services/utils";
+import { updateTicket } from "../Services";
+import { getCurrentDate, agentSelectionDicio } from "../../utils";
 import dbg from "../../dbg";
-import { updateSolidarityTickets } from "../../graphql/mutations";
 
 const log = dbg.extend("updateIndividualTicket");
 
@@ -40,9 +35,13 @@ const hasuraSchema = yup
   })
   .required();
 
-export default async (ticket_id: number, state: string, agent: number) => {
+export default async (
+  { ticket_id, state }: { ticket_id: number; state: string },
+  agent: number
+) => {
   log("Couldn't find any close volunteers for MSR");
   const ticket = {
+    ticket_id,
     status: "pending",
     assignee_id: agentSelectionDicio[agent],
     custom_fields: [
@@ -65,38 +64,5 @@ export default async (ticket_id: number, state: string, agent: number) => {
       public: false
     }
   };
-  try {
-    log(`Updating MSR ticket '${ticket_id}' in Zendesk...`);
-    const zendeskTicket = await updateTicket(ticket_id, ticket);
-    if (!zendeskTicket) {
-      throw new Error("Zendesk ticket update returned errors");
-    }
-
-    log(`Preparing ticket '${zendeskTicket.id}' to be saved in Hasura`);
-    const hasuraTicket = {
-      ...zendeskTicket,
-      ...composeCustomFields(zendeskTicket.custom_fields),
-      ticket_id: zendeskTicket.id
-    };
-
-    const validatedTicket = await hasuraSchema.validate(hasuraTicket, {
-      stripUnknown: true
-    });
-
-    log(
-      `Updating individual ticket '${validatedTicket.ticket_id}' in Hasura...`
-    );
-    const inserted = await updateSolidarityTickets(validatedTicket, [
-      validatedTicket.ticket_id
-    ]);
-    if (!inserted)
-      log(
-        `Something went wrong when updating this MSR ticket in Hasura '${zendeskTicket.id}'`
-      );
-
-    return zendeskTicket.id;
-  } catch (e) {
-    log("failed to create ticket in zendesk: ".red, e);
-    return undefined;
-  }
+  return await updateTicket(ticket, hasuraSchema);
 };
